4.21 Deleting a blog is now limited to the blog's author

+++ src/controllers/blogsControllers.js
@@ -8,12 +8,12 @@ async function getAllBlogs(req, res) {
 
 async function postBlog(req, res) {
     const { title, url, likes=0 } = req.body
-    const author = req.user;
+    const author = req.token;
     if(!title || !author || !url ){
         return res.status(400).send('Missing information');
     }
-    const user = await User.findOne({username: author});
-    const postResponse = await Blog.create({ title, author, url, likes, user: user.id });
+    const user = await User.findOne({_id: author});
+    const postResponse = await Blog.create({ title, author: user.username, url, likes, user: user.id });
     user.blogs = user.blogs.concat(postResponse.id);
     await user.save();
     res.status(201).json(postResponse);
@@ -22,6 +22,10 @@ async function postBlog(req, res) {
 async function deleteBlog(req, res) {
     const { id } = req.params;
     try{
+        const blog = await Blog.findOne({ _id: id });
+        if(blog.user.toString() !== req.token){
+            return res.status(403).send("Can't delete a blog that isn't yours");
+        }
         await Blog.deleteOne({_id: id});
         return res.status(200).send('Deleted successfully');
     } catch(error) {
